More unit tests for item-list.component and item-matcher-offer.component
