+++ src/app/components/item-list/item-list.component.spec.ts
@@ -1,13 +1,15 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterOutletStubComponent, RouterLinkStubDirective, RouterStub, ActivatedRouteStub, ActivatedRoute, Router } from '../../../test/router-stubs';
 
+import { Item } from '../../classes/pojo/item';
+import { ItemListComponent } from './item-list.component';
+import { ItemService } from '../../services/item.service';
+import { ItemServiceMock } from '../../../test/item-service-mock';
 import { LoadingComponent } from '../loading/loading.component';
 import { MessageComponent } from '../message/message.component';
-import { ItemListComponent } from './item-list.component';
 import { PaginationComponent } from '../pagination/pagination.component';
-import { ItemService } from '../../services/item.service';
 import { SearchResult } from '../../classes/search/search-result';
-import { Item } from '../../classes/pojo/item';
+import { Pagination } from '../../classes/search/pagination';
 
 const activatedRouteMock = {
 	snapshot: {
@@ -15,12 +17,6 @@ const activatedRouteMock = {
 			get: function (a: any) { return undefined }
 		}
 	}
-}
-
-class ItemServiceMock {
-  search(){
-    return new Promise<SearchResult<Item>>(() => {});
-  };
 }
 
 describe('ItemListComponent', () => {
@@ -54,7 +50,26 @@ describe('ItemListComponent', () => {
 		fixture.detectChanges();
 	});
 
-	it('should create', () => {
+	it('should display items list', () => {
+		component.ngOnInit();
+		fixture.whenStable().then(() => {
+			fixture.detectChanges();
+			expect(fixture.nativeElement.querySelectorAll("td").length).toBe(2);
+			expect(fixture.nativeElement.querySelector(".mt-control-panel button")).toBeDefined();			
+		});
 		expect(component).toBeTruthy();
 	});
+
+	it('should display [no items found message] when there are no items', () => {
+    let injectedItemService = fixture.debugElement.injector.get(ItemService);
+    spyOn(injectedItemService, 'search').and.callFake((a) =>{
+			const searchResult = new SearchResult<Item>(new Array<Item>(), new Pagination(1,10,0));
+      return Promise.resolve(searchResult);
+		});
+		component.ngOnInit();
+		fixture.whenStable().then(() => {
+			fixture.detectChanges();
+			expect(fixture.nativeElement.querySelector('.items-not-found')).toBeDefined();
+		});
+	});	
 });
